chore(root): cofigure eslint
