+++ .eslintrc.json
@@ -1,0 +1,7 @@
+{
+    "extends": "airbnb",
+    "parser": "babel-eslint",
+    "rules": {
+        "strict": 0
+    }
+}
