+++ src/pages/documentation.jsx
@@ -49,7 +49,9 @@ class DocumentationPage extends React.Component {
   onClickComponent = (name, componentData) => {
     const { location } = this.props;
     navigate(`/documentation?component=${name.toLowerCase()}`);
-    window.gtag('config', 'UA-41862404-1', { page_path: `${location.pathname}/${name.toLowerCase()}` });
+    if (window) {
+      window.gtag('config', 'UA-41862404-1', { page_path: `${location.pathname}/${name.toLowerCase()}` });
+    }
     this.setState({
       activeComponent: name,
       componentData,
