+++ src/pages/usage.jsx
@@ -49,7 +49,9 @@ class UsagePage extends React.Component {
   onClickComponent = (name, componentData) => {
     const { location } = this.props;
     navigate(`/usage?component=${name.toLowerCase()}`);
-    window.gtag('config', 'UA-41862404-1', { page_path: `${location.pathname}/${name.toLowerCase()}` });
+    if (window) {
+      window.gtag('config', 'UA-41862404-1', { page_path: `${location.pathname}/${name.toLowerCase()}` });
+    }
     this.setState({
       activeComponent: name,
       componentData,
