+++ src/templates/landing-main/contrib-section/footer.jsx
