+++ src/templates/landing-main/index.jsx
@@ -1,11 +1,11 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react';
 import { Link } from 'gatsby';
+import Button from 'react-lite-ui/components/button';
 import iphone from '../../images/i-phone-6.svg';
 import ipad from '../../images/i-pad-air.svg';
 import desktop from '../../images/desktop.svg';
 import FeaturesSection from './features';
-import Button from 'react-lite-ui/components/button';
 import ContributionSection from './contrib-section';
 
 import './index.scss';
